requirements.txt
asgiref==3.4.1
astroid==2.9.0
Django==3.2.6
djangorestframework==3.12.4
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==4.0.0
wrapt==1.13.3
yapf==0.31.0
django-mysql==4.5.0